Revert "SERVER-43762 tighten the overload set for BSONObjBuilder::appendNumber"
This reverts commit 3307ba57eef2dfff9abe2413e39058ac9451ec04.

+++ src/mongo/db/commands/generic_servers.cpp
@@ -279,8 +279,7 @@ public:
             }
             typename RamLogType::LineIterator rl(ramlog);
 
-            result.appendNumber("totalLinesWritten",
-                                static_cast<long long>(rl.getTotalLinesWritten()));
+            result.appendNumber("totalLinesWritten", rl.getTotalLinesWritten());
 
             BSONArrayBuilder arr(result.subarrayStart("log"));
             while (rl.more())
